Update: Error Message
In this commit, the error messages for the input field somewhat works.

The error messages show when clicking "Stem nu".

It works correctly when there is no input in the field. However, for the other error messages to show after clicking the "Stem nu", you need to remove the input in order for the error message to show.